requirements.txt
requests
beautifulsoup4
pdfkit
lxml
langchain
langchain-community
langchain-experimental
langsmith
langgraph
llama-index
llama-parse
fastembed
chromadb
python-dotenv
langchain-groq
chainlit
fastembed
unstructured[md]
python-dotenv
python-telegram-bot